requirements.txt
autopep8==1.6.0
boto3==1.18.53
botocore==1.21.53
filesplit==3.0.2
jmespath==0.10.0
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
